confluence_md_exporter/requirements.txt
confluence-markdown-exporter>=3.0.0
python-dotenv>=1.0.0
